+++ crates/opt8n/src/cli.rs
@@ -3,7 +3,6 @@ use anvil::cmd::NodeArgs;
 use clap::{Command, CommandFactory, Parser, Subcommand};
 use color_eyre::eyre;
 use forge_script::ScriptArgs;
-use serde::{Deserialize, Serialize};
 use std::fmt::Debug;
 use tracing::trace;
 
@@ -64,21 +63,4 @@ impl Cli {
             arg
         })
     }
-}
-
-#[derive(Parser, Clone, Debug, Serialize, Deserialize, PartialEq, Eq)]
-#[clap(rename_all = "kebab_case", infer_subcommands = true, multicall = true)]
-pub enum Opt8nCommand {
-    #[command(visible_alias = "a")]
-    Anvil {
-        #[arg(index = 1, allow_hyphen_values = true)]
-        args: Vec<String>,
-    },
-    #[command(visible_alias = "c")]
-    Cast {
-        #[arg(index = 1, allow_hyphen_values = true)]
-        args: Vec<String>,
-    },
-    #[command(visible_alias = "e")]
-    Exit,
 }
