+++ crates/opt8n/src/main.rs
@@ -1,5 +1,6 @@
 pub mod cli;
 pub mod opt8n;
+pub mod repl;
 
 use clap::FromArgMatches;
 use color_eyre::eyre;
